Tema2/Activitat5/requirements.txt
plotly~=5.18.0
mysql-connector-python~=8.2.0